refactor: move name formatter to util package
