+++ packages/svrx-util/lib/svrx-util.js
@@ -1,3 +1,4 @@
+exports.nameFormatter = require('./name-formatter');
 exports.rcFileRead = require('./rc-file-read');
 exports.logger = require('./logger');
 exports.npm = require('./npm');
