+++ packages/svrx/lib/util/helper.js
@@ -5,10 +5,7 @@ const libPath = require('path');
 const libFs = require('fs');
 const _ = require('lodash');
 const os = require('os');
-const CONST = require('../constant');
 
-const { PLUGIN_PREFIX } = CONST;
-const scopeAndNameRegex = /^@([a-z\d][\w-.]+)\/([a-z\d][\w-.]*)/;
 const o2str = {}.toString;
 const { slice } = [];
 
@@ -24,52 +21,6 @@ function npCall(callback, args, ctx) {
 
     callback.apply(ctx, args);
   });
-}
-
-/**
- * combine pluginName to packageName
- * - foo -> svrx-plugin-foo
- * - foo-bar -> svrx-plugin-foo-bar
- * - @scope/foo -> @scope/svrx-plugin-foo
- * - @scope/foo-bar -> @scope/svrx-plugin-foo-bar
- * @param name
- * @returns {string|*}
- */
-function normalizePluginName(name) {
-  const combineName = (n) => (n.startsWith(PLUGIN_PREFIX) ? n : PLUGIN_PREFIX + n);
-  const isScoped = name.startsWith('@');
-
-  if (isScoped) {
-    const matches = scopeAndNameRegex.exec(name);
-    if (matches) {
-      const scope = matches[1];
-      const realName = matches[2];
-      return `@${scope}/${combineName(realName)}`;
-    }
-    return null;
-  }
-  return combineName(name);
-}
-
-/**
- * parse packageName to pluginName (revert normalizePluginName())
- * @param packageName
- */
-function parsePluginName(packageName) {
-  const isScoped = packageName.startsWith('@');
-  const removePrefix = (n) => (n.startsWith(PLUGIN_PREFIX) ? n.slice(PLUGIN_PREFIX.length) : null);
-
-  if (isScoped) {
-    const matches = scopeAndNameRegex.exec(packageName);
-    if (matches) {
-      const scope = matches[1];
-      const realName = matches[2];
-      const formattedName = removePrefix(realName);
-      return formattedName ? `@${scope}/${formattedName}` : null;
-    }
-    return null;
-  }
-  return removePrefix(packageName);
 }
 
 function isReadableStream(test) {
@@ -194,8 +145,6 @@ function isFn(o) {
 }
 
 
-exports.normalizePluginName = normalizePluginName;
-exports.parsePluginName = parsePluginName;
 exports.getExternalIp = _.memoize(getExternalIp);
 exports.isReadableStream = isReadableStream;
 exports.getCert = _.memoize(getCert);
